WIP for #3
Some experiments with putting the final target URL of the OIDC dance into a query parameter of the OIDC callback URL.

This definitely requires an IdP that allows for query params in callback URLs. So, if this ever proves to be useful, it has to be tamed behind a flag.